Extract some shared code between FSD versions into helper class
