+++ Inferences/ForwardSubsumptionDemodulation2.cpp
@@ -1,4 +1,5 @@
 #include "ForwardSubsumptionDemodulation2.hpp"
+#include "SubsumptionDemodulationHelper.hpp"
 
 #include "Debug/RuntimeStatistics.hpp"
 #include "Indexing/Index.hpp"
@@ -489,138 +490,6 @@ std::ostream& operator<<(std::ostream& o, OverlayBinder const& binder)
 }
 
 
-#if VDEBUG
-/// Returns true iff clause with literal lits1 is smaller than clause with literals lits2
-/// in the multiset extension of the given ordering.
-///
-/// This implementation is justified by Lemma 2.5.6 on page 24 of [BN98].
-/// [BN98] Franz Baader and Tobias Nipkow. Term Rewriting and All That. Cambridge University Press, 1998.
-bool clauseIsSmaller(Literal* const lits1[], unsigned n1, Literal* const lits2[], unsigned n2, Ordering const& ordering)
-{
-  // Copy given literals so we can sort them
-  v_vector<Literal*> c1(lits1, lits1+n1);
-  v_vector<Literal*> c2(lits2, lits2+n2);
-
-  // These will contain literals from c1/c2 with equal occurrences removed
-  v_vector<Literal*> v1;
-  v_vector<Literal*> v2;
-
-  // Sort input by pointer value
-  // NOTE: we use std::less<> because the C++ standard guarantees it is a total order on pointer types.
-  //       (the built-in operator< is not required to be a total order for pointer types.)
-  std::less<Literal*> const lit_ptr_less;
-  std::sort(c1.begin(), c1.end(), lit_ptr_less);
-  std::sort(c2.begin(), c2.end(), lit_ptr_less);
-
-  // Skip occurrences of equal literals
-  unsigned i1 = 0;
-  unsigned i2 = 0;
-  while (i1 < n1 && i2 < n2) {
-    if (c1[i1] == c2[i2]) {
-      // skip this occurrence
-      ++i1;
-      ++i2;
-    }
-    else if (lit_ptr_less(c1[i1], c2[i2])) {
-      v1.push_back(c1[i1]);
-      ++i1;
-    }
-    else if (lit_ptr_less(c2[i2], c1[i1])) {
-      v2.push_back(c2[i2]);
-      ++i2;
-    }
-    else {
-      ASSERTION_VIOLATION;
-    }
-  }
-  while (i1 < n1) {
-    ASS_GE(i2, n2);
-    v1.push_back(c1[i1]);
-    ++i1;
-  }
-  while (i2 < n2) {
-    ASS_GE(i1, n1);
-    v2.push_back(c2[i2]);
-    ++i2;
-  }
-
-  if (v1.empty() && v2.empty()) {
-    // Both clauses are the same
-    ASS(c1 == c2);
-    return false;
-  }
-
-  // For each remaining literal from c1,
-  // we have to find a greater one in the remaining ones from c2.
-  for (Literal* l1 : v1) {
-    bool isCovered = false;
-    for (Literal* l2 : v2) {
-      switch (ordering.compare(l1, l2)) {
-        case Ordering::LESS:
-          // yay
-          isCovered = true;
-          break;
-        case Ordering::INCOMPARABLE:
-        case Ordering::GREATER:
-          // doesn't work
-          break;
-        case Ordering::EQUAL:
-          // should not happen due to first part where we remove equal literals
-          ASSERTION_VIOLATION;
-        case Ordering::LESS_EQ:
-        case Ordering::GREATER_EQ:
-          // those don't appear
-          ASSERTION_VIOLATION;
-        default:
-          ASSERTION_VIOLATION;
-      }
-      if (isCovered) {
-        break;
-      }
-    }
-    if (!isCovered) {
-      return false;
-    }
-  }
-
-  return true;
-}
-#endif  // VDEBUG
-
-
-/**
- * Build clause that results from subsumption resolution with main premise 'cl' and side premise 'mcl'.
- * The literal 'resLit' is the resolved literal from 'cl'.
- */
-Clause* generateSubsumptionResolutionClause(Clause* cl, Literal* resLit, Clause* mcl)
-{
-  CALL("generateSubsumptionResolutionClause");
-
-  Inference* inference = new Inference2(Inference::SUBSUMPTION_RESOLUTION, cl, mcl);
-  Unit::InputType inputType = std::max(cl->inputType(), mcl->inputType());
-
-  unsigned newLen = cl->length() - 1;
-  Clause* newCl = new(newLen) Clause(newLen, inputType, inference);
-
-  unsigned j = 0;
-  for (unsigned i = 0; i < cl->length(); ++i) {
-    Literal* curLit = (*cl)[i];
-
-    if (curLit != resLit) {
-      (*newCl)[j] = curLit;
-      j += 1;
-    }
-  }
-  // We should have skipped exactly one literal, namely resLit.
-  // (it should never appear twice because we apply duplicate literal removal before subsumption resolution)
-  ASS_EQ(j, newLen);
-
-  newCl->setAge(cl->age());
-
-  return newCl;
-}
-
-
 }  // namespace
 
 
@@ -679,7 +548,7 @@ bool ForwardSubsumptionDemodulation2::perform(Clause* cl, Clause*& replacement, 
   // Initialize miniIndex with literals in the clause cl
   // TODO(idea for later): maybe it helps to order alternatives, either smaller to larger or larger to smaller, or unordered
   // to do this, we can simply order the literals inside the miniIndex (i.e., in each equivalence class w.r.t. literal header)
-  LiteralMiniIndex const miniIndex(cl);
+  LiteralMiniIndex const cl_miniIndex(cl);
 
   unsigned int const cl_maxVar = cl->maxVar();
 
@@ -724,7 +593,7 @@ bool ForwardSubsumptionDemodulation2::perform(Clause* cl, Clause*& replacement, 
       for (unsigned mi = 0; mi < mcl->length(); ++mi) {
         Literal* baseLit = (*mcl)[mi];
 
-        LiteralMiniIndex::InstanceIterator instIt(miniIndex, baseLit, false);
+        LiteralMiniIndex::InstanceIterator instIt(cl_miniIndex, baseLit, false);
 
         if (!instIt.hasNext()) {
           // baseLit does not have any suitable alternative at all!
@@ -1098,8 +967,8 @@ bool ForwardSubsumptionDemodulation2::perform(Clause* cl, Clause*& replacement, 
                     // Here, we have subsumption resolution
                     ASS_EQ(binder.applyTo(eqLit), Literal::complementaryLiteral(dlit));  // ¬eqLitS == dlit
                     ASS_EQ(ordering.compare(binder.applyTo(eqLit), dlit), Ordering::GREATER);  // L > ¬L
-                    // TODO: ASS(checkForSubsumptionResolution(cl, cm, dlit));
-                    replacement = generateSubsumptionResolutionClause(cl, dlit, mcl);
+                    ASS(SDHelper::checkForSubsumptionResolution(cl, ClauseMatches{mcl,cl_miniIndex}, dlit));
+                    replacement = SDHelper::generateSubsumptionResolutionClause(cl, dlit, mcl);
                     premises = pvi(getSingletonIterator(mcl));
                     env.statistics->forwardSubsumptionResolution++;
                     return true;
@@ -1198,7 +1067,7 @@ isRedundant:
               if (getOptions().literalComparisonMode() != Options::LiteralComparisonMode::REVERSE) {  // see note above
                 // Check newCl < cl.
                 // This is quite obvious, there should be no problems with this.
-                ASS(clauseIsSmaller(newCl->literals(), newCl->length(), cl->literals(), cl->length(), ordering));
+                ASS(SDHelper::clauseIsSmaller(newCl->literals(), newCl->length(), cl->literals(), cl->length(), ordering));
                 // Check mclΘ < cl.
                 // This is not so clear and might easily be violated if we have a bug above.
                 v_vector<Literal*> mclS(mcl->literals(), mcl->literals() + mcl->length());
@@ -1206,7 +1075,7 @@ isRedundant:
                 for (auto it = mclS.begin(); it != mclS.end(); ++it) {
                   *it = binder.applyTo(*it);
                 }
-                if (!clauseIsSmaller(mclS.data(), mclS.size(), cl->literals(), cl->length(), ordering)) {
+                if (!SDHelper::clauseIsSmaller(mclS.data(), mclS.size(), cl->literals(), cl->length(), ordering)) {
                   std::cerr << "FSDv2: redundancy violated!" << std::endl;
                   std::cerr << "mcl: " << mcl->toString() << std::endl;
                   std::cerr << " cl: " <<  cl->toString() << std::endl;
