+++ Inferences/ForwardSubsumptionDemodulation3.cpp
@@ -1,4 +1,5 @@
 #include "ForwardSubsumptionDemodulation3.hpp"
+#include "SubsumptionDemodulationHelper.hpp"
 
 #include "Debug/RuntimeStatistics.hpp"
 #include "Indexing/Index.hpp"
@@ -8,7 +9,6 @@
 #include "Kernel/ColorHelper.hpp"
 #include "Kernel/EqHelper.hpp"
 #include "Kernel/Inference.hpp"
-#include "Kernel/MLMatcher.hpp"
 #include "Kernel/MLMatcher2.hpp"
 #include "Kernel/Matcher.hpp"
 #include "Kernel/Ordering.hpp"
@@ -260,279 +260,6 @@ std::ostream& operator<<(std::ostream& o, OverlayBinder const& binder)
 }
 
 
-#if VDEBUG
-/// Returns true iff clause with literal lits1 is smaller than clause with literals lits2
-/// in the multiset extension of the given ordering.
-///
-/// This implementation is justified by Lemma 2.5.6 on page 24 of [BN98].
-/// [BN98] Franz Baader and Tobias Nipkow. Term Rewriting and All That. Cambridge University Press, 1998.
-bool clauseIsSmaller(Literal* const lits1[], unsigned n1, Literal* const lits2[], unsigned n2, Ordering const& ordering)
-{
-  // Copy given literals so we can sort them
-  v_vector<Literal*> c1(lits1, lits1+n1);
-  v_vector<Literal*> c2(lits2, lits2+n2);
-
-  // These will contain literals from c1/c2 with equal occurrences removed
-  v_vector<Literal*> v1;
-  v_vector<Literal*> v2;
-
-  // Sort input by pointer value
-  // NOTE: we use std::less<> because the C++ standard guarantees it is a total order on pointer types.
-  //       (the built-in operator< is not required to be a total order for pointer types.)
-  std::less<Literal*> const lit_ptr_less;
-  std::sort(c1.begin(), c1.end(), lit_ptr_less);
-  std::sort(c2.begin(), c2.end(), lit_ptr_less);
-
-  // Skip occurrences of equal literals
-  unsigned i1 = 0;
-  unsigned i2 = 0;
-  while (i1 < n1 && i2 < n2) {
-    if (c1[i1] == c2[i2]) {
-      // skip this occurrence
-      ++i1;
-      ++i2;
-    }
-    else if (lit_ptr_less(c1[i1], c2[i2])) {
-      v1.push_back(c1[i1]);
-      ++i1;
-    }
-    else if (lit_ptr_less(c2[i2], c1[i1])) {
-      v2.push_back(c2[i2]);
-      ++i2;
-    }
-    else {
-      ASSERTION_VIOLATION;
-    }
-  }
-  while (i1 < n1) {
-    ASS_GE(i2, n2);
-    v1.push_back(c1[i1]);
-    ++i1;
-  }
-  while (i2 < n2) {
-    ASS_GE(i1, n1);
-    v2.push_back(c2[i2]);
-    ++i2;
-  }
-
-  if (v1.empty() && v2.empty()) {
-    // Both clauses are the same
-    ASS(c1 == c2);
-    return false;
-  }
-
-  // For each remaining literal from c1,
-  // we have to find a greater one in the remaining ones from c2.
-  for (Literal* l1 : v1) {
-    bool isCovered = false;
-    for (Literal* l2 : v2) {
-      switch (ordering.compare(l1, l2)) {
-        case Ordering::LESS:
-          // yay
-          isCovered = true;
-          break;
-        case Ordering::INCOMPARABLE:
-        case Ordering::GREATER:
-          // doesn't work
-          break;
-        case Ordering::EQUAL:
-          // should not happen due to first part where we remove equal literals
-          ASSERTION_VIOLATION;
-        case Ordering::LESS_EQ:
-        case Ordering::GREATER_EQ:
-          // those don't appear
-          ASSERTION_VIOLATION;
-        default:
-          ASSERTION_VIOLATION;
-      }
-      if (isCovered) {
-        break;
-      }
-    }
-    if (!isCovered) {
-      return false;
-    }
-  }
-
-  return true;
-}
-#endif  // VDEBUG
-
-
-/**
- * Stores an instance of the multi-literal matching problem.
- * Allows us to re-use the alts from subsumption for subsumption resolution.
- */
-class ClauseMatches
-{
-  public:
-    ClauseMatches(Clause* base, LiteralMiniIndex const& ixAlts);
-
-    ~ClauseMatches();
-
-    // Disallow copy
-    ClauseMatches(ClauseMatches const&) = delete;
-    ClauseMatches& operator=(ClauseMatches const&) = delete;
-
-    // Default move is fine
-    ClauseMatches(ClauseMatches&&) = default;
-    ClauseMatches& operator=(ClauseMatches&&) = default;
-
-    Clause* base() const { return m_base; }
-    LiteralList const* const* alts() const { return m_alts.data(); }
-
-    unsigned baseLitsWithoutAlts() const { return m_baseLitsWithoutAlts; }
-
-    bool isSubsumptionPossible() const
-    {
-      // For subsumption, every base literal must have at least one alternative
-      return m_baseLitsWithoutAlts == 0;
-    }
-
-    bool isSubsumptionDemodulationPossible() const
-    {
-      ASS_GE(m_baseLitsWithoutAlts, m_basePosEqsWithoutAlts);
-      // Demodulation needs at least one positive equality
-      if (m_basePosEqs == 0) {
-        return false;
-      }
-      // If there are base literals without any suitable alternatives:
-      // 1. If there is only one literal without alternative and it is a positive equality,
-      //    then it might still be possible to get an FSD inference by choosing this literal
-      //    as equality for demodulation.
-      // 2. If there is a literal without alternative but it is not a positive equality,
-      //    then it is impossible to get an FSD inference.
-      // 3. If there are two literals without alternatives, then it is impossible as well.
-      return m_baseLitsWithoutAlts == 0  // every base literal has alternatives
-        || (m_baseLitsWithoutAlts == 1 && m_basePosEqsWithoutAlts == 1);  // case 1 in comment above
-    }
-
-  private:
-    Clause* m_base;
-    v_vector<LiteralList*> m_alts;
-    unsigned m_basePosEqs;
-    unsigned m_baseLitsWithoutAlts;
-    unsigned m_basePosEqsWithoutAlts;
-};  // class ClauseMatches
-
-ClauseMatches::ClauseMatches(Clause* base, LiteralMiniIndex const& ixAlts)
-  : m_base(base)
-  , m_alts(base->length(), LiteralList::empty())
-  , m_basePosEqs(0)
-  , m_baseLitsWithoutAlts(0)
-  , m_basePosEqsWithoutAlts(0)
-{
-  for (unsigned i = 0; i < m_base->length(); ++i) {
-    Literal* baseLit = (*m_base)[i];
-    bool isPosEq = baseLit->isEquality() || baseLit->isPositive();
-
-    if (isPosEq) {
-      m_basePosEqs += 1;
-    }
-
-    LiteralMiniIndex::InstanceIterator instIt(ixAlts, baseLit, false);
-
-    if (!instIt.hasNext()) {
-      // baseLit does not have any suitable alternative at all!
-      m_baseLitsWithoutAlts += 1;
-      if (isPosEq) {
-        m_basePosEqsWithoutAlts += 1;
-      }
-    }
-
-    ASS(LiteralList::isEmpty(m_alts[i]));
-    while (instIt.hasNext()) {
-      Literal* matched = instIt.next();
-      LiteralList::push(matched, m_alts[i]);
-    }
-  }
-}
-
-ClauseMatches::~ClauseMatches()
-{
-  for (LiteralList* ll : m_alts) {
-    LiteralList::destroy(ll);
-  }
-}
-
-
-/**
- * Check whether there is a subsumption resolution inference with main premise 'cl'
- * using resolved literal 'resLit' (from 'cl'), and side premise represented by 'cm'.
- */
-bool checkForSubsumptionResolution(Clause* cl, ClauseMatches const& cm, Literal* resLit)
-{
-  Clause* mcl = cm.base();
-
-  if (cm.baseLitsWithoutAlts() > 0) {
-    // Got base literals without alts?
-    // Then subsumption resolution is possible,
-    // but only if we can resolve ALL of them away
-    // (i.e., complementary-match with resLit)
-    //
-    // NOTE: if there is more than one base literal without alts,
-    // subsumption resolution might still be possible
-    // (if we can unify them and resolve them all away)
-    for (unsigned i = 0; i < mcl->length(); ++i) {
-      if (LiteralList::isEmpty(cm.alts()[i])) {
-        if (!MatchingUtils::match((*mcl)[i], resLit, /* complementary = */ true)) {
-          return false;
-        }
-      }
-    }
-
-  } else {
-    // No base literal without alts?
-    // Subsumption resolution is possible if at least one base lit can be resolved with resLit
-
-    bool anyResolvable = false;
-    for (unsigned i = 0; i < mcl->length(); ++i) {
-      if (MatchingUtils::match((*mcl)[i], resLit, /* complementary = */ true)) {
-        anyResolvable = true;
-        break;
-      }
-    }
-    if (!anyResolvable) {
-      return false;
-    }
-  }
-
-  // NOTE: we use MLMatcher here because we want *subset* inclusion (as opposed to submultiset)
-  return MLMatcher::canBeMatched(mcl, cl, cm.alts(), resLit);
-}
-
-
-/**
- * Build clause that results from subsumption resolution with main premise 'cl' and side premise 'mcl'.
- * The literal 'resLit' is the resolved literal from 'cl'.
- */
-Clause* generateSubsumptionResolutionClause(Clause* cl, Literal* resLit, Clause* mcl)
-{
-  CALL("generateSubsumptionResolutionClause");
-
-  Inference* inference = new Inference2(Inference::SUBSUMPTION_RESOLUTION, cl, mcl);
-  Unit::InputType inputType = std::max(cl->inputType(), mcl->inputType());
-
-  unsigned newLen = cl->length() - 1;
-  Clause* newCl = new(newLen) Clause(newLen, inputType, inference);
-
-  unsigned j = 0;
-  for (unsigned i = 0; i < cl->length(); ++i) {
-    Literal* curLit = (*cl)[i];
-
-    if (curLit != resLit) {
-      (*newCl)[j] = curLit;
-      j += 1;
-    }
-  }
-  // We should have skipped exactly one literal, namely resLit.
-  // (it should never appear twice because we apply duplicate literal removal before subsumption resolution)
-  ASS_EQ(j, newLen);
-
-  newCl->setAge(cl->age());
-
-  return newCl;
-}
 
 
 }  // namespace
@@ -1000,8 +727,8 @@ bool ForwardSubsumptionDemodulation3::perform(Clause* cl, Clause*& replacement, 
                     // Here, we have subsumption resolution
                     ASS_EQ(binder.applyTo(eqLit), Literal::complementaryLiteral(dlit));  // ¬eqLitS == dlit
                     ASS_EQ(ordering.compare(binder.applyTo(eqLit), dlit), Ordering::GREATER);  // L > ¬L
-                    ASS(checkForSubsumptionResolution(cl, cm, dlit));
-                    replacement = generateSubsumptionResolutionClause(cl, dlit, mcl);
+                    ASS(SDHelper::checkForSubsumptionResolution(cl, cm, dlit));
+                    replacement = SDHelper::generateSubsumptionResolutionClause(cl, dlit, mcl);
                     premises = pvi(getSingletonIterator(mcl));
                     env.statistics->forwardSubsumptionResolution++;
                     return true;
@@ -1100,7 +827,7 @@ isRedundant:
               if (getOptions().literalComparisonMode() != Options::LiteralComparisonMode::REVERSE) {  // see note above
                 // Check newCl < cl.
                 // This is quite obvious, there should be no problems with this.
-                ASS(clauseIsSmaller(newCl->literals(), newCl->length(), cl->literals(), cl->length(), ordering));
+                ASS(SDHelper::clauseIsSmaller(newCl->literals(), newCl->length(), cl->literals(), cl->length(), ordering));
                 // Check mclΘ < cl.
                 // This is not so clear and might easily be violated if we have a bug above.
                 v_vector<Literal*> mclS(mcl->literals(), mcl->literals() + mcl->length());
@@ -1108,7 +835,7 @@ isRedundant:
                 for (auto it = mclS.begin(); it != mclS.end(); ++it) {
                   *it = binder.applyTo(*it);
                 }
-                if (!clauseIsSmaller(mclS.data(), mclS.size(), cl->literals(), cl->length(), ordering)) {
+                if (!SDHelper::clauseIsSmaller(mclS.data(), mclS.size(), cl->literals(), cl->length(), ordering)) {
                   std::cerr << "FSDv3: redundancy violated!" << std::endl;
                   std::cerr << "mcl: " << mcl->toString() << std::endl;
                   std::cerr << " cl: " <<  cl->toString() << std::endl;
@@ -1144,7 +871,7 @@ isRedundant:
       while (rit.hasNext()) {
         Clause* mcl = rit.next().clause;
         if (ColorHelper::compatible(cl->color(), mcl->color())) {
-          Clause* resCl = generateSubsumptionResolutionClause(cl, resLit, mcl);
+          Clause* resCl = SDHelper::generateSubsumptionResolutionClause(cl, resLit, mcl);
           env.statistics->forwardSubsumptionResolution++;
           premises = pvi(getSingletonIterator(mcl));
           replacement = resCl;
@@ -1158,8 +885,8 @@ isRedundant:
       Clause* mcl = cm.base();
       for (unsigned li = 0; li < cl->length(); ++li) {
         Literal* resLit = (*cl)[li];  // resolved literal
-        if (ColorHelper::compatible(cl->color(), mcl->color()) && checkForSubsumptionResolution(cl, cm, resLit)) {
-          Clause* resCl = generateSubsumptionResolutionClause(cl, resLit, mcl);
+        if (ColorHelper::compatible(cl->color(), mcl->color()) && SDHelper::checkForSubsumptionResolution(cl, cm, resLit)) {
+          Clause* resCl = SDHelper::generateSubsumptionResolutionClause(cl, resLit, mcl);
           env.statistics->forwardSubsumptionResolution++;
           premises = pvi(getSingletonIterator(mcl));
           replacement = resCl;
@@ -1193,8 +920,8 @@ isRedundant:
         ClauseMatches const& cm = altsStorage.back();
         ASS_EQ(cm.base(), mcl);  // make sure we got the right one (since C++17, emplace_back returns the new element)
 
-        if (ColorHelper::compatible(cl->color(), mcl->color()) && checkForSubsumptionResolution(cl, cm, resLit)) {
-          Clause* resCl = generateSubsumptionResolutionClause(cl, resLit, mcl);
+        if (ColorHelper::compatible(cl->color(), mcl->color()) && SDHelper::checkForSubsumptionResolution(cl, cm, resLit)) {
+          Clause* resCl = SDHelper::generateSubsumptionResolutionClause(cl, resLit, mcl);
           env.statistics->forwardSubsumptionResolution++;
           premises = pvi(getSingletonIterator(mcl));
           replacement = resCl;
